+++ src/PayPal/Api/Plan/BillingCycle/Frequency.php
@@ -1,0 +1,50 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Swag\PayPal\PayPal\Api\Plan\BillingCycle;
+
+use Swag\PayPal\PayPal\Api\Common\PayPalStruct;
+
+/**
+ * @codeCoverageIgnore
+ * @experimental
+ *
+ * This class is experimental and not officially supported.
+ * It is currently not used within the plugin itself. Use with caution.
+ */
+class Frequency extends PayPalStruct
+{
+    /**
+     * @var string
+     */
+    protected $intervalUnit;
+
+    /**
+     * @var int
+     */
+    protected $intervalCount;
+
+    public function getIntervalUnit(): string
+    {
+        return $this->intervalUnit;
+    }
+
+    public function setIntervalUnit(string $intervalUnit): void
+    {
+        $this->intervalUnit = $intervalUnit;
+    }
+
+    public function getIntervalCount(): int
+    {
+        return $this->intervalCount;
+    }
+
+    public function setIntervalCount(int $intervalCount): void
+    {
+        $this->intervalCount = $intervalCount;
+    }
+}