+++ src/PayPal/Api/Plan/PaymentPreferences.php
@@ -1,0 +1,50 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Swag\PayPal\PayPal\Api\Plan;
+
+use Swag\PayPal\PayPal\Api\Common\PayPalStruct;
+
+/**
+ * @codeCoverageIgnore
+ * @experimental
+ *
+ * This class is experimental and not officially supported.
+ * It is currently not used within the plugin itself. Use with caution.
+ */
+class PaymentPreferences extends PayPalStruct
+{
+    /**
+     * @var bool
+     */
+    protected $autoBillOutstanding;
+
+    /**
+     * @var int
+     */
+    protected $paymentFailureThreshold;
+
+    public function getAutoBillOutstanding(): bool
+    {
+        return $this->autoBillOutstanding;
+    }
+
+    public function setAutoBillOutstanding(bool $autoBillOutstanding): void
+    {
+        $this->autoBillOutstanding = $autoBillOutstanding;
+    }
+
+    public function getPaymentFailureThreshold(): int
+    {
+        return $this->paymentFailureThreshold;
+    }
+
+    public function setPaymentFailureThreshold(int $paymentFailureThreshold): void
+    {
+        $this->paymentFailureThreshold = $paymentFailureThreshold;
+    }
+}