+++ src/PayPal/Api/Subscription/BillingInfo/LastPayment/Amount.php
@@ -1,0 +1,21 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Swag\PayPal\PayPal\Api\Subscription\BillingInfo\LastPayment;
+
+use Swag\PayPal\PayPal\Api\Common\Money;
+
+/**
+ * @codeCoverageIgnore
+ * @experimental
+ *
+ * This class is experimental and not officially supported.
+ * It is currently not used within the plugin itself. Use with caution.
+ */
+class Amount extends Money
+{
+}