+++ src/PayPal/Api/Subscription/Subscriber/Name.php
@@ -1,0 +1,50 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Swag\PayPal\PayPal\Api\Subscription\Subscriber;
+
+use Swag\PayPal\PayPal\Api\Common\PayPalStruct;
+
+/**
+ * @codeCoverageIgnore
+ * @experimental
+ *
+ * This class is experimental and not officially supported.
+ * It is currently not used within the plugin itself. Use with caution.
+ */
+class Name extends PayPalStruct
+{
+    /**
+     * @var string
+     */
+    protected $givenName;
+
+    /**
+     * @var string
+     */
+    protected $surname;
+
+    public function getGivenName(): string
+    {
+        return $this->givenName;
+    }
+
+    public function setGivenName(string $givenName): void
+    {
+        $this->givenName = $givenName;
+    }
+
+    public function getSurname(): string
+    {
+        return $this->surname;
+    }
+
+    public function setSurname(string $surname): void
+    {
+        $this->surname = $surname;
+    }
+}